requirements.txt
amqp==5.0.6
aniso8601==9.0.1
attrs==21.2.0
billiard==3.6.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.2
flask-expects-json==1.7.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
kombu==5.2.2
MarkupSafe==2.0.1
prompt-toolkit==3.0.23
pyrsistent==0.18.0
pyserial==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
